lesson2 assignment is done
